Time: 0 ms (100.00%), Space: 43.9 MB (5.32%) - LeetHub

+++ 162-find-peak-element/162-find-peak-element.java
@@ -10,7 +10,7 @@ class Solution {
        {
         return mid;
        }
-if(lval<(long)arr[mid])
+if(lval<arr[mid])
 {
     low=mid+1;
 }else
